MiuiCamera: Disable burst shot and QRCode for now
QRCode is disabled by default even on stock. It's based on
xiaomi's qrcode reader, that is dependent on MIUI's framework. I'll check
later if we can port it to aosp.

Burst shot crashes the camera on custom roms, it can only be fixed by rom
devs. (Anyone interested, check here: https://github.com/Magisk-Modules-Repo/Xiaomi-Mi-A1-Stock-Camera/issues/5 )

Also, this commit removes the tissot_sprout.xml from inside the apk, as
it makes the one at system/etc/ useless. (and it's easier to debug if it's
outside the apk... Maybe later we can revert this, for a self-contained apk)